requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
faiss-cpu
requests
